refactor PetTooltip component and update import path
