+++ src/layouts/widgetify-card/pets/components/pet-tooltip.tsx
